Added rudimentary theme support
Theme support is NOW! Woo! The documentation is nonexistent, and some
of the details are sketchy, but IT WORKS!

Included is the Base16 Monokai theme. Later to come might include
Zenburn and others! Yay!

This required some massive restructuring of how Messages are handled,
and how their colors are managed. In this case, a new Sub-Enum is created,
the Message.MessageColor enum, which defines a bunch of possible values.

Later on, I need to document all these.