New feature (slash fix to a bug): Reconnecting on transient network
Network changes now schedule a reconnect. This has added a new permission (network state).
The old feature of having a reconnecting network will most likely go away soon, as it is pretty much
rendundant (this also saves batteries, as it uses broadcasts instead of
a wakelock on the system clock)