Partial fix for #40
So, there's a lot of things that went into this. The ties between model.Conversation and a ConversationAdapter are negligible. As a result, I've done some things I aint proud of. It works, but it isn't at all clean.

Further addition: Understanding the "*" meta-channel. It's a thought, but down the road.