Rizon (and other servers) should work better-er now.
This hopefully fixes #55. It works for me: I can connect to Rizon, and things work fine.

So what's going on here is that yaaic had some serious state problems. _inputThread inside PircBot apparently handles sending bytes to the server as well as fetching lines FROM the server.

sendRawLine inside pIRCbot now just checks if _inputThread or _outputThread are null, making sure that if for some reason it's all broken that we don't do anything really mean and bad. As a result, I now explicitly call in places that need to make sure that _inputThread pumps things out, it's explicitly called.

OutputThread now flushes the IO before sending a line *as well as* after sending the line. This should clear up having any network problems.

IRCService apparently now has a newline stashed away in there, too. Take it as a sign of my insanity.