Clean up message rendering
I'm not sure this is the right way to do this; I'd like to clean up
the number of places I do string operations and avoid doing that.

I am considering making themes not affect the history,
instead making the background of each message the background of the
theme they were rendered with.