it's 03h05 and I don't want to write a damn commit message
So here it goes:

I fixed things. A lot of things. In a lot of cases, I've fixed bugs that are
important for the current space. For example, AddAliasActivity is now a
LinearLayout and is really a dialog container.

I need to make it a real dialog. Just like Pinnochio needs to be a real boy
and get laid by a real woman not a cheap $5 hooker.

This commit message is nasty. Dear future self: don't write commit
messages at 03h10.

There's some proofing in App.java that makes it less likely to blow up in
your face. There's some cool shit I've mangled in CommandListView
that makes it less likely to blow up in your face.

I've muddled with some strings. Some words got changerized. Indicator
nubs got fiddled with.

Shoot me and the horse I rode in on. Please, deliver me from this.

I've been staring at this code for the last week or so and committing
in batches. This commit message is the final one before I say
"fukkit" and mangle some words, make some words up and add
a tag and release this sonnava to the world.

To whoever reads this commit message: Thank you for getting this far.
You deserve a beer. I need a stiff drink. There's a few other things
I could go for but those requests don't belong in a commit message.

There's some debug slime let over in ConversationActivty. It's mostly the
record of me trying to hunt down something going very bad. This is okay
though, as the failure mode described within that WTF condition is
perfectly valid. If I ever see it in a log anywhere, I'm going to flip doops.

Load this sucker in your mainframe, because it's one ✇ of a mother✇ker.

I'm going to bed. Goodnight world.

Signed-off-by: Morgan Gangwere <morgan.gangwere@gmail.com>